Added tests. Change to BigDecimal.

+++ src/main/java/es/davidmartos/model/Symbol.java
@@ -1,17 +1,18 @@
 package es.davidmartos.model;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
+import java.math.BigDecimal;
 import lombok.Data;
 
 @Data
 public class Symbol {
 
   @JsonProperty("reward_multiplier")
-  private Double rewardMultiplier;
+  private BigDecimal rewardMultiplier;
 
   private String type;
 
   private String impact;
 
-  private Integer extra;
+  private BigDecimal extra;
 }
